Uploading portfolio project from school
This was a project that I worked on for school for my portfolio. However, that repository was privately owned by the school and I was not able to make it public. I created this repository so it may be public to future employers, but unfortunately, there is not as much commit history.